Refactor PostStatusTask from MainActivity
