+++ app/src/main/java/edu/byu/cs/tweeter/client/presenter/MainPresenter.java
@@ -1,5 +1,12 @@
 package edu.byu.cs.tweeter.client.presenter;
 
+import java.net.MalformedURLException;
+import java.text.ParseException;
+import java.text.SimpleDateFormat;
+import java.time.LocalDate;
+import java.time.LocalTime;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 
@@ -10,6 +17,8 @@ import edu.byu.cs.tweeter.client.model.service.StatusService;
 import edu.byu.cs.tweeter.client.model.service.UserService;
 import edu.byu.cs.tweeter.client.model.service.backgroundTask.IsFollowerTask;
 import edu.byu.cs.tweeter.client.model.service.backgroundTask.LogoutTask;
+import edu.byu.cs.tweeter.client.model.service.backgroundTask.PostStatusTask;
+import edu.byu.cs.tweeter.model.domain.Status;
 import edu.byu.cs.tweeter.model.domain.User;
 
 public class MainPresenter {
@@ -19,6 +28,7 @@ public class MainPresenter {
         void setCount(boolean isFollowingCount, int count);
         void updateFollowButton(boolean removed);
         void enableFollowButton(boolean value);
+        void successfulPost();
         void logoutUser();
     }
 
@@ -62,6 +72,77 @@ public class MainPresenter {
 
     public void onLogoutButtonClicked() {
         userService.logout(Cache.getInstance().getCurrUserAuthToken(), new LogoutObserver());
+    }
+
+    public void onStatusPosted(String post) throws ParseException, MalformedURLException {
+        Status newStatus = new Status(post, Cache.getInstance().getCurrUser(), getFormattedDateTime(),
+                                        parseURLs(post), parseMentions(post));
+        statusService.postStatus(Cache.getInstance().getCurrUserAuthToken(), newStatus,
+                                new PostStatusObserver());
+    }
+
+    private String getFormattedDateTime() throws ParseException {
+        SimpleDateFormat userFormat = new SimpleDateFormat("yyyy-MM-dd hh:mm:ss");
+        SimpleDateFormat statusFormat = new SimpleDateFormat("MMM d yyyy h:mm aaa");
+
+        return statusFormat.format(userFormat.parse(LocalDate.now().toString() + " " + LocalTime.now().toString().substring(0, 8)));
+    }
+
+    private List<String> parseURLs(String post) throws MalformedURLException {
+        List<String> containedUrls = new ArrayList<>();
+        for (String word : post.split("\\s")) {
+            if (word.startsWith("http://") || word.startsWith("https://")) {
+
+                int index = findUrlEndIndex(word);
+
+                word = word.substring(0, index);
+
+                containedUrls.add(word);
+            }
+        }
+
+        return containedUrls;
+    }
+
+    private List<String> parseMentions(String post) {
+        List<String> containedMentions = new ArrayList<>();
+
+        for (String word : post.split("\\s")) {
+            if (word.startsWith("@")) {
+                word = word.replaceAll("[^a-zA-Z0-9]", "");
+                word = "@".concat(word);
+
+                containedMentions.add(word);
+            }
+        }
+
+        return containedMentions;
+    }
+
+    private int findUrlEndIndex(String word) {
+        if (word.contains(".com")) {
+            int index = word.indexOf(".com");
+            index += 4;
+            return index;
+        } else if (word.contains(".org")) {
+            int index = word.indexOf(".org");
+            index += 4;
+            return index;
+        } else if (word.contains(".edu")) {
+            int index = word.indexOf(".edu");
+            index += 4;
+            return index;
+        } else if (word.contains(".net")) {
+            int index = word.indexOf(".net");
+            index += 4;
+            return index;
+        } else if (word.contains(".mil")) {
+            int index = word.indexOf(".mil");
+            index += 4;
+            return index;
+        } else {
+            return word.length();
+        }
     }
 
     public class GetFollowersCountObserver implements FollowService.GetFollowersCountObserver {
@@ -183,6 +264,23 @@ public class MainPresenter {
         @Override
         public void handleException(Exception ex) {
             view.displayMessage("Failed to logout because of exception: " + ex.getMessage());
+        }
+    }
+
+    public class PostStatusObserver implements StatusService.PostStatusObserver {
+        @Override
+        public void handleSuccess() {
+            view.successfulPost();
+        }
+
+        @Override
+        public void handleFailure(String message) {
+            view.displayMessage("Failed to post status: " + message);
+        }
+
+        @Override
+        public void handleException(Exception ex) {
+            view.displayMessage("Failed to post status because of exception: " + ex.getMessage());
         }
     }
 }
